==> script/2. Data Cleansing.ipynb <==
# Data Cleansing
---

<div class="alert alert-block alert-danger">
<b>Warning:</b> Do not change this script
</div>

import pandas as pd
from tqdm.notebook import tqdm

path = "/home/jupyter/agra/source code/sentiment analytics/"

df = pd.read_csv("data/resto-scrape.csv")

# head data
df.head()

## Remove Duplicate

# count duplicate
total_duplicates = df.duplicated(subset=["text"]).sum()
print("total duplicates : ", total_duplicates)

# remove duplicates
print("before remove : ", len(df))

df = df.drop_duplicates(subset=["text"], keep="last")

print("after remove : ", len(df))

## Remove Spam

# check spam tweet
idx = []
for i in tqdm(range(df.shape[0])):
    if i == (df.shape[0] - 1):
        break
    temp_a = df.iloc[i].text.split()[:7]
    # print(df.iloc[i].text) # display text
    temp_b = df.iloc[i+1].text.split()[:7]
    # print(df.iloc[i+1].text) # display text
    if temp_a == temp_b:
        idx.append(i+1)

# total spam tweet

print("total spam tweet : ", len(idx))

df.to_csv("resto-cleaned.csv", index=False)
