==> script/3. Data Preprocessing & Labelling.ipynb <==
# Data Preprocessing
---

<div class="alert alert-block alert-danger">
<b>Warning:</b> Do not change this script
</div>

import pandas as pd
import swifter
import re

from tqdm.notebook import tqdm

import nltk
from nltk.stem import WordNetLemmatizer
from nltk.corpus import stopwords
from nltk.tokenize import word_tokenize
# nltk.download('punkt') --> uncomment if you haven't downloaded it yet
# nltk.download('wordnet') --> uncomment if you haven't downloaded it yet

from Sastrawi.StopWordRemover.StopWordRemoverFactory import StopWordRemoverFactory
from Sastrawi.Stemmer.StemmerFactory import StemmerFactory

path = "/home/jupyter/agra/source code/sentiment analytics/"

df = pd.read_csv(path + "data/resto-cleaned.csv")

df = df[df['language'] == 'in']

cols = ["text"]
df = df[cols]

## Preparation
---

def filtering(text):
    text = re.sub('[0-9]+', '', str(text)) #removing numberic value
    text = re.sub(r'#', '', text) #removing '#' symbol
    text = re.sub(r'[\n]+', '', text) # remove new line
    text = re.sub(r"^\s+|\s+$", "", text) #remove leading and trailing spaces in a word using OR sign to delete both
    text = re.sub(r" +", " ", text) #remove multiple space betwen words
    text = re.sub('https? :\/\/\S+', '', text) #removing hyperlink / URL
    text = re.sub(r"\b[a-zAZ]\b", "", text) #removing single char
    text = re.sub('\s+',' ',text) #removing multiple whitespace
    text = text.replace('\\t'," ").replace('\\n'," ").replace('\\u'," ").replace('\\',"") #remove tab, new line, and back slice
    text = re.sub(r'[^\w\s]', '', text) #remove puntuation& emoji (remove all besides \w > word dan \s > space)
    text = re.sub(r'(.)1+', r'1', text) # remove repeating character
    text = re.sub("[^a-zA-Z]",' ',text)
    
    return text

def casefoldingText(text): # Converting all the characters in a text into lower case
    text = text.lower() 
    return text

def tokenize(text) :
    text = nltk.word_tokenize(text)
    return text

with open(path + "dict/stopwords1.txt") as f:
    additional_stopwords1 = [line.strip() for line in f]

with open(path + "dict/stopwords2.txt") as f:
    additional_stopwords2 = [line.strip() for line in f]

with open(path + "dict/number-stopword.txt") as f:
    number_stopword = [line.strip() for line in f]
    
with open(path + "dict/calendar-words.txt") as f:
    calendar_stopword = [line.strip() for line in f]
    
with open(path + "dict/indonesian-region.txt") as f:
    region_stopword = [line.strip() for line in f]

with open(path + "dict/swear-words.txt") as f:
    swear_stopword = [line.strip() for line in f]
    
with open(path + "dict/resto-words.txt") as f:
    resto_stopword = [line.strip() for line in f]
    
def stopword(text): # Remove stopwors in a text
    listStopwords = set(stopwords.words('indonesian'))
    filtered = []
    for txt in text:
        if txt not in listStopwords:
            filtered.append(txt)
    text = filtered 
    return text

def Slangword(ulasan):
    kamusSlang = eval(open(path+"dict/combined_slang_words.txt").read())
    pattern = re.compile(r'\b( ' + '|'.join (kamusSlang.keys())+r')\b')
    content = []
    for kata in ulasan:
        filterSlang = pattern.sub(lambda x: kamusSlang[x.group()],kata)
        content.append(filterSlang.lower())
    ulasan = content
    return ulasan

def stopwords(words):
    return [word for word in words if word not in data]

def stemmed_wrapper(term):
    return stemmer.stem(term)

def stemmingText(document):
    return [term_dict[term] for term in document]

def check_exaggeration(text):
    """Mengidentifikasi apakah terdapat pengulangan huruf yang berlebihan dalam teks"""
    count = 0
    for i, char in enumerate(text):
        if i < len(text) - 1 and char == text[i + 1]:
            count += 1
    if count > len(text) * 0.3:
        return True
    else:
        return False
    
def repair_exaggeration(text):
    """Menghilangkan pengulangan huruf yang berlebihan dalam teks"""
    new_text = ""
    for i, char in enumerate(text):
        if i < len(text) - 1 and char == text[i + 1]:
            continue
        new_text += char
    return new_text

## Start Preprocessing
---

### Filtering

%time df['filtering'] = df['text'].swifter.apply(filtering)

print('before filtering : ', df['text'].iloc[1])
print('after filtering : ', df['filtering'].iloc[1])

### Casefolding

%time df['casefolding'] = df['filtering'].swifter.apply(casefoldingText)

print('before casefolding : ', df['filtering'].iloc[1])
print('after casefolding : ', df['casefolding'].iloc[1])

### Repair Exaggeration

%time df['check_exaggeration'] = df['casefolding'].swifter.apply(check_exaggeration)
check_exaggeration_df = df[df['check_exaggeration'] == True]

print("total exaggeration word : ", len(check_exaggeration_df))

%time df['repair_exaggeration'] = df['casefolding'].swifter.apply(repair_exaggeration)

print('before repair_exaggeration :\n', df['casefolding'].iloc[3318],'\n')
print('after repair_exaggeration :\n', df['repair_exaggeration'].iloc[3318])

stop_factory = StopWordRemoverFactory()

# custom your stopword
more_stopword = additional_stopwords

data = stop_factory.get_stop_words()+more_stopword
stopword = stop_factory.create_stop_word_remover()

%time df['stopwords'] = df['repair_exaggeration'].swifter.apply(stopwords)

print('before stopwords :\n', df['repair_exaggeration'].iloc[2494],'\n')
print('after stopwords :\n', df['stopwords'].iloc[2494])

### Tokenize

%time df['tokenize'] = df['repair_exaggeration'].swifter.apply(tokenize)

print('before tokenize :\n', df['repair_exaggeration'].iloc[1],'\n')
print('after tokenize :\n', df['tokenize'].iloc[1])

### Stopword

stop_factory = StopWordRemoverFactory()

# custom your stopword
more_stopword = ['jir','njir','bjir','juga', 'yang','yg', 'nya', 'aja', 'loh' ,'sih', 'deh', 'an', 'ro', 
                 'aj', 'kopi','kenangan','tanyarlfes','convomfs','FOODFESS2','FOODFESS', 'hehe', 'lho',
                'fodfes', 'worksfes', 'discountfes', 'bankneocommerce']

data = stop_factory.get_stop_words()+more_stopword+additional_stopwords1+additional_stopwords2+number_stopword+calendar_stopword+region_stopword+swear_stopword+resto_stopword
stopword = stop_factory.create_stop_word_remover()

%time df['stopword'] = df['tokenize'].swifter.apply(stopwords)

print('before stopword :\n', df['tokenize'].iloc[1],'\n')
print('after stopword :\n', df['stopword'].iloc[1])

### Formalization

%time df['formalisasi'] = df['stopword'].swifter.apply(Slangword)

print('before formalisasi :\n', df['stopword'].iloc[1],'\n')
print('after formalisasi :\n', df['formalisasi'].iloc[1])

### Stemming

factory = StemmerFactory()
stemmer = factory.create_stemmer()

term_dict = {}

for document in df['formalisasi']:
    for term in document:
        if term not in term_dict:
            term_dict[term] = ' '

for term in tqdm(term_dict): term_dict[term] = stemmed_wrapper(term)
    # print(term,":" ,term_dict[term]) --> display text

%time df['stemming'] = df['formalisasi'].swifter.apply(stemmingText)

print('after stemming :\n', df['formalisasi'].iloc[1],'\n')
print('before stemming :\n', df['stemming'].iloc[1])

### Result

df.head()

df.to_csv(path + "data/resto-preprocess-2.csv", index=False)

# Data Labelling
---

<div class="alert alert-block alert-danger">
<b>Warning:</b> Do not change this script
</div>

## Preparation
---

# Loads lexicon positive and negative data
lexicon_positive = dict()
import csv
with open(path + 'dict/lexicon_positive.csv', 'r') as csvfile:
    reader = csv.reader(csvfile, delimiter=',')
    for row in reader:
        lexicon_positive[row[0]] = int(row[1])

lexicon_negative = dict()
import csv
with open(path + 'dict/lexicon_negative.csv', 'r') as csvfile:
    reader = csv.reader(csvfile, delimiter=',')
    for row in reader:
        lexicon_negative[row[0]] = int(row[1])
        
# Function to determine sentiment polarity of tweets        
def sentiment_analysis_lexicon_indonesia(text):
    #for word in text:
    score = 0
    for word in text:
        if (word in lexicon_positive):
            score = score + lexicon_positive[word]
    for word in text:
        if (word in lexicon_negative):
            score = score + lexicon_negative[word]
    polarity=''
    if (score > 0):
        polarity = 'positive'
    elif (score < 0):
        polarity = 'negative'
    else:
        polarity = 'neutral'
    return score, polarity

## Start Labelling using Lexicon
---

%time results = df['stemming'].swifter.apply(sentiment_analysis_lexicon_indonesia)
results = list(zip(*results))

df['polarity_score'] = results[0]
df['polarity'] = results[1]
print(df['polarity'].value_counts())

### Result

df.to_csv(path + "data/resto-labelling-2.csv", index=False)

df.head()

## Comparasion Sentiment Polarity on Tweets Data

pd.set_option('display.max_colwidth', 3000)
positive_tweets = df[df['polarity'] == 'positive']
positive_tweets = positive_tweets[['text', 'polarity_score', 'polarity']].sort_values(by = 'polarity_score', ascending=False).reset_index(drop = True)
positive_tweets.index += 1
positive_tweets[0:10]

pd.set_option('display.max_colwidth', 3000)
negative_tweets = df[df['polarity'] == 'negative']
negative_tweets = negative_tweets[['text', 'polarity_score', 'polarity']].sort_values(by = 'polarity_score', ascending=True)[0:10].reset_index(drop = True)
negative_tweets.index += 1
negative_tweets[0:10]

pd.set_option('display.max_colwidth', 3000)
neutral_tweets = df[df['polarity'] == 'neutral']
neutral_tweets = neutral_tweets[['text', 'polarity_score', 'polarity']].sort_values(by = 'polarity_score', ascending=True)[0:10].reset_index(drop = True)
neutral_tweets.index += 1
neutral_tweets[0:10]
