==> script/4. Feature Selection.ipynb <==
# Feature Selection
---

<div class="alert alert-block alert-danger">
<b>Warning:</b> Do not change this script
</div>

# import libraries
import pandas as pd
import re

path = "/home/jupyter/agra/source code/sentiment analytics/"

df = pd.read_csv(path+"data/resto-labelling-2.csv")

df.head()

### Preparation

def filtering(text):
    text = re.sub('[0-9]+', '', str(text)) #removing numberic value
    text = re.sub(r'#', '', text) #removing '#' symbol
    text = re.sub(r'[\n]+', '', text) # remove new line
    text = re.sub(r"^\s+|\s+$", "", text) #remove leading and trailing spaces in a word using OR sign to delete both
    text = re.sub(r" +", " ", text) #remove multiple space betwen words
    text = re.sub('https? :\/\/\S+', '', text) #removing hyperlink / URL
    text = re.sub(r"\b[a-zAZ]\b", "", text) #removing single char
    text = re.sub('\s+',' ',text) #removing multiple whitespace
    text = text.replace('\\t'," ").replace('\\n'," ").replace('\\u'," ").replace('\\',"") #remove tab, new line, and back slice
    text = re.sub(r'[^\w\s]', '', text) #remove puntuation& emoji (remove all besides \w > word dan \s > space)
    text = re.sub(r'(.)1+', r'1', text) # remove repeating character
    text = re.sub("[^a-zA-Z]",' ',text)
    
    return text

## Start Feature Engineering
---

%time df['final_text'] = df['stemming'].apply(filtering)

cols = ["text","polarity_score"]
df = df[cols]

df["degree_of_polarity"] = None
for i, row in df.iterrows():
    if row["polarity_score"] > 0:
        df.loc[i, "degree_of_polarity"] = 2
    elif row["polarity_score"] < 0:
        df.loc[i, "degree_of_polarity"] = 0
    else:
        df.loc[i, "degree_of_polarity"] = 1

cols = ["text","degree_of_polarity"]
df = df[cols]

df

## Simple Analysis

pd.set_option('display.max_colwidth', 3000)
positive_tweet = df[df['degree_of_polarity'] == 2]
positive_tweet = positive_tweet[['final_text', 'degree_of_polarity']].reset_index(drop = True)
# positive_tweet.index += 1
positive_tweet[20:40]

pd.set_option('display.max_colwidth', 3000)
negative_tweets = df[df['degree_of_polarity'] == 0]
negative_tweets = negative_tweets[['final_text', 'degree_of_polarity']].reset_index(drop = True)
negative_tweets.index += 1
negative_tweets[0:20]

df.to_csv(path + "data/resto-feature-engineering-final2.csv", index=False, header=False)
